KU-118/refactor: /types/response.ts에 API 응답 타입 분리
